@@ -85,10 +85,10 @@ class Stats extends Component {
   }
 
   componentDidMount() {
-    const { id } = this.props;
+    const { domain, id } = this.props;
     if (!id) return null;
     return axios
-      .get(`/api/url/stats?id=${id}`, { headers: { Authorization: cookie.get('token') } })
+      .get(`/api/url/stats?id=${id}&domain=${domain}`, { headers: { Authorization: cookie.get('token') } })
       .then(({ data }) =>
         this.setState({
           stats: data,
@@ -126,8 +126,8 @@ class Stats extends Component {
         <TitleWrapper>
           <Title>
             Stats for:{' '}
-            <a href={stats.shortUrl} title="Short URL">
-              {stats.shortUrl.replace(/https?:\/\//, '')}
+            <a href={stats.shortLink} title="Short link">
+              {stats.shortLink.replace(/https?:\/\//, '')}
             </a>
           </Title>
           <TitleTarget>
@@ -155,6 +155,7 @@ class Stats extends Component {
 
 Stats.propTypes = {
   isAuthenticated: PropTypes.bool.isRequired,
+  domain: PropTypes.string.isRequired,
   id: PropTypes.string.isRequired,
   showPageLoading: PropTypes.func.isRequired,
 };
